modified:   Aulas/aula10.py 	modified:   "Exercic\303\255os/Aula 10/ex29.py"

+++ Exercicíos/Aula 10/ex29.py
@@ -1,5 +1,5 @@
 limite_velocidade = 50
-velocidade_carro = int(input('Digite a velocidade que o carro passou no radar: '))
+velocidade_carro = float(input('Digite a velocidade que o carro passou no radar: '))
 multa = (velocidade_carro - 50) * 7
 
 if velocidade_carro < limite_velocidade:
